fix type mismatch

+++ src/mumu.h
@@ -54,7 +54,7 @@ struct Parameters {
   std::string log;
 
   // default values
-  unsigned int threads {threads_default};
+  unsigned long int threads {threads_default};
   double minimum_match {minimum_match_default};
   std::string minimum_ratio_type {use_minimum_value};
   double minimum_ratio {minimum_ratio_default};
